+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0-preview] - 2020-01-13
+
+### Changed
+* Updated `com.unity.platforms` package to version `0.2.0-preview`.
+* Updated `com.unity.platforms.desktop` package to version `0.2.0-preview`.
+
+## [0.1.8-preview] - 2019-12-11
+
+### Changed
+* Bump com.unity.platforms version to 0.1.8-preview
+
 ## [0.1.7-preview.4] - 2019-12-09
 
 ### Changed